requirements.txt
appnope==0.1.4
asttokens==3.0.1
comm==0.2.3
contourpy==1.3.3
cycler==0.12.1
debugpy==1.8.19
decorator==5.2.1
executing==2.2.1
fonttools==4.61.1
graphviz==0.21
ipykernel==7.1.0
ipython==9.9.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
matplotlib==3.10.8
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
networkx==3.6.1
numpy==2.4.1
packaging==26.0
parso==0.8.5
pexpect==4.9.0
pillow==12.1.0
platformdirs==4.5.1
prompt_toolkit==3.0.52
psutil==7.2.2
ptyprocess==0.7.0
pure_eval==0.2.3
Pygments==2.19.2
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pyzmq==27.1.0
scipy==1.17.0
six==1.17.0
stack-data==0.6.3
tornado==6.5.4
traitlets==5.14.3
wcwidth==0.5.0
